Fix implicit conversion error
Fixes the following compile error when building on Windows with
GN arg use_custom_libcxx=false.

../../base/allocator/partition_allocator/src\partition_alloc/thread_cache.h(187,42): error: implicit conversion loses integer precision: 'uint16_t' (aka 'unsigned short') to 'uint8_t' (aka 'unsigned char') [-Werror,-Wimplicit-int-conversion]
  187 |   uint8_t largest_active_bucket_index_ = internal::BucketIndexLookup::GetIndex(
      |                                          ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  188 |       ThreadCacheLimits::kDefaultSizeThreshold);
      |       ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
.

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5120035
Commit-Queue: Marshall Greenblatt <marshall@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1237115}

+++ base/allocator/partition_allocator/src/partition_alloc/thread_cache.cc
@@ -206,7 +206,7 @@ void ThreadCacheRegistry::ForcePurgeAllThreadAfterForkUnsafe() {
 }
 
 void ThreadCacheRegistry::SetLargestActiveBucketIndex(
-    uint8_t largest_active_bucket_index) {
+    uint16_t largest_active_bucket_index) {
   largest_active_bucket_index_ = largest_active_bucket_index;
 }
